Add files via upload
This is a project on movie recommendation system which will recommend you the movies to watch. This is my  Practice projects. I have used this project as my practice project .The source of this project is from kaggle. The reference for this project :https://www.kaggle.com/code/ibtesama/getting-started-with-a-movie-recommendation-system 
LINK FOR DATA SET :https://drive.google.com/drive/folders/115ZN8cv6a4OBNuOOsVOjMhsGDFlHT8Yu?usp=share_link